# shared/RLG Final Regional Complete file/app.js
/**
 * app.js - Main backend API for RLG Data & RLG Fans
 *
 * This application delivers:
 *   - AI-driven data insights & predictive analytics
 *   - Real-time scraping & market intelligence
 *   - Compliance monitoring and report generation
 *   - Dynamic, geolocation-based pricing with region-specific tiers:
 *       * Special Region Pricing for Israel ("עם ישראל חי!, הפתרון הטכנולוגי שישנה את העתיד.")
 *       * SADC Region Pricing for select African countries
 *       * Global Default Pricing
 *   - Monetization strategies, newsletter distribution, and agent chatbot interactions
 *   - Integrated RLG Super Tool for actionable insights
 *   - Secure user authentication and system health checks
 *
 * Swagger (OpenAPI 3.0) documentation is auto‑generated and available at /docs and /redoc.
 */

const express = require('express');
const bodyParser = require('body-parser');
const swaggerUi = require('swagger-ui-express');
const swaggerJSDoc = require('swagger-jsdoc');
const cors = require('cors');

// Create Express application
const app = express();

// Enable CORS, JSON body parsing, etc.
app.use(cors());
app.use(bodyParser.json());
app.use(bodyParser.urlencoded({ extended: true }));

// ---------------------------------------------------------------------
// Swagger Definitions (OpenAPI 3.0)
// ---------------------------------------------------------------------
const swaggerDefinition = {
  openapi: '3.0.0',
  info: {
    title: 'RLG Data & RLG Fans API',
    version: '1.0.0',
    description:
      'A comprehensive AI-driven platform for data insights, real-time scraping, compliance monitoring, dynamic geolocation-based pricing (including Special Region pricing for Israel and SADC tiers), monetization strategies, reporting, newsletter distribution, agent chat bot, and RLG Super Tool integration.',
  },
  servers: [
    {
      url: 'http://localhost:8000',
      description: 'Local Development Server',
    },
    {
      url: 'https://api.rlgmedia.com',
      description: 'Production Server',
    },
  ],
};

const options = {
  swaggerDefinition,
  apis: ['./app.js'], // This file; add other files as needed.
};

const swaggerSpec = swaggerJSDoc(options);
app.use('/docs', swaggerUi.serve, swaggerUi.setup(swaggerSpec));

// ---------------------------------------------------------------------
// Stub Implementations for Core Functionalities
// In production, these should be replaced with actual modules/services.
// ---------------------------------------------------------------------

// AI Analysis stub
function runAIAnalysis(dataFile, targetColumn, featuresColumns) {
  // Real implementation should process the file and return insights.
  return { insights: 'Analysis results for ' + dataFile };
}

// Scraping stub
function runScraping(url, keywords) {
  return { url: url, keywords: keywords, data: "Scraped content from " + url };
}

// Compliance check stub
function runComplianceCheck(mediaId) {
  return { media_id: mediaId, status: "approved", notes: "Compliant with GDPR and CCPA" };
}

// RLG Super Tool stub
function getSuperToolInsights() {
  return { insights: "Advanced insights from RLG Super Tool" };
}

// Authentication stub (dummy user store; replace with secure DB & hashing)
const usersDB = {
  admin: {
    username: 'admin',
    password: 'password', // DO NOT store plain-text passwords in production!
    location: { country: 'Israel', city: 'Tel Aviv' },
    pricing: { special: true },
    special_message: 'עם ישראל חי!, הפתרון הטכנולוגי שישנה את העתיד.'
  },
  user456: {
    username: 'user456',
    password: 'password456',
    location: { country: 'United States', city: 'New York' },
    pricing: { special: false },
    special_message: ''
  }
};

// Report generation stub
function generateReport(reportType) {
  return { report_type: reportType, data: "Report content for " + reportType };
}

// Newsletter stub
function sendNewsletter(content) {
  return { status: "sent", content: content };
}

// Chat bot stub
function processAgentChat(message) {
  return { reply: "Automated reply to: " + message };
}

// Pricing endpoint: Dynamic, geolocation-based pricing logic.
// For demonstration we use IP address lookup stubs; real implementation would integrate with a geolocation service.
function getPricing(userIp) {
  // For simplicity, assume that if userIp includes 'IL' then it is Israel.
  // In a real scenario, use an API (like ipapi.co) to determine country.
  const ipString = userIp || "";
  if (ipString.includes("IL")) {
    return {
      region: "Special Region",
      pricing: {
        Creator: { weekly: 35, monthly: 99 },
        Pro: { weekly: 65, monthly: 199 },
        Enterprise: { monthly: 699 },
        "RLG Media Pack": { monthly: 2500 }
      },
      special_message: "עם ישראל חי!, הפתרון הטכנולוגי שישנה את העתיד."
    };
  }
  // SADC sample stub; in production check against a full list.
  if (ipString.includes("ZA") || ipString.includes("BW")) {
    return {
      region: "Africa & SADC",
      pricing: {
        Creator: { weekly: 12, monthly: 49 },
        Pro: { weekly: 25, monthly: 79 },
        Enterprise: { monthly: 269 },
        "RLG Media Pack": { monthly: 1400 }
      },
      special_message: ""
    };
  }
  // Default Global Pricing
  return {
    region: "Global",
    pricing: {
      Creator: { weekly: 15, monthly: 59 },
      Pro: { weekly: 35, monthly: 99 },
      Enterprise: { monthly: 599 },
      "RLG Media Pack": { monthly: 2000 }
    },
    special_message: ""
  };
}

// ---------------------------------------------------------------------
// API Endpoints
// ---------------------------------------------------------------------

/**
 * @swagger
 * /:
 *   get:
 *     tags:
 *       - Root
 *     summary: Returns a welcome message.
 *     responses:
 *       200:
 *         description: Welcome message with platform overview.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 message:
 *                   type: string
 *                   example: "Welcome to RLG Data & RLG Fans API!"
 *                 description:
 *                   type: string
 *                   example: "This platform provides AI-driven insights, real-time scraping, compliance monitoring, dynamic pricing and more."
 */
app.get('/', (req, res) => {
  res.json({
    message: "Welcome to RLG Data & RLG Fans API!",
    description: "This platform provides AI-driven insights, real-time scraping, compliance monitoring, dynamic geolocation-based pricing, monetization strategies, reporting, newsletter distribution, and an integrated agent chat bot. Access docs at /docs or /redoc."
  });
});

/**
 * @swagger
 * /ai_analysis:
 *   post:
 *     tags:
 *       - AI Analysis
 *     summary: Execute AI analysis pipeline.
 *     requestBody:
 *       description: Parameters for AI analysis.
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             $ref: '#/components/schemas/AIAnalysisRequest'
 *     responses:
 *       200:
 *         description: Analysis executed successfully.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 status:
 *                   type: string
 *                   example: "success"
 *                 result:
 *                   type: object
 *       500:
 *         description: AI analysis failure.
 */
app.post('/ai_analysis', (req, res) => {
  try {
    const { data_file, target_column, features_columns } = req.body;
    const result = runAIAnalysis(data_file, target_column, features_columns);
    res.json({ status: "success", result: result });
  } catch (error) {
    logger.error("AI analysis failed: " + error.stack);
    res.status(500).json({ detail: "AI analysis failed: " + error.message });
  }
});

/**
 * @swagger
 * /scrape:
 *   post:
 *     tags:
 *       - Scraping
 *     summary: Initiate a scraping job.
 *     requestBody:
 *       description: Scraping parameters.
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             $ref: '#/components/schemas/ScrapeRequest'
 *     responses:
 *       200:
 *         description: Scraping executed successfully.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 status:
 *                   type: string
 *                   example: "success"
 *                 data:
 *                   type: object
 *       500:
 *         description: Scraping failure.
 */
app.post('/scrape', (req, res) => {
  try {
    const { url, keywords } = req.body;
    const data = runScraping(url, keywords);
    res.json({ status: "success", data: data });
  } catch (error) {
    logger.error("Scraping failed: " + error.stack);
    res.status(500).json({ detail: "Scraping failed: " + error.message });
  }
});

/**
 * @swagger
 * /compliance:
 *   post:
 *     tags:
 *       - Compliance
 *     summary: Perform a compliance check on a media record.
 *     requestBody:
 *       description: Media ID for compliance check.
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               media_id:
 *                 type: integer
 *                 example: 123
 *     responses:
 *       200:
 *         description: Compliance check successful.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 status:
 *                   type: string
 *                   example: "success"
 *                 compliance:
 *                   type: object
 *       500:
 *         description: Compliance check failed.
 */
app.post('/compliance', (req, res) => {
  try {
    const { media_id } = req.body;
    const compliance = runComplianceCheck(media_id);
    res.json({ status: "success", compliance: compliance });
  } catch (error) {
    logger.error("Compliance check failed: " + error.stack);
    res.status(500).json({ detail: "Compliance check failed: " + error.message });
  }
});

/**
 * @swagger
 * /super_tool:
 *   get:
 *     tags:
 *       - RLG Super Tool
 *     summary: Retrieve RLG Super Tool insights.
 *     responses:
 *       200:
 *         description: Insights retrieved successfully.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 status:
 *                   type: string
 *                   example: "success"
 *                 insights:
 *                   type: object
 *       500:
 *         description: Super Tool failure.
 */
app.get('/super_tool', (req, res) => {
  try {
    const insights = getSuperToolInsights();
    res.json({ status: "success", insights: insights });
  } catch (error) {
    logger.error("RLG Super Tool failed: " + error.stack);
    res.status(500).json({ detail: "Super Tool failed: " + error.message });
  }
});

/**
 * @swagger
 * /login:
 *   post:
 *     tags:
 *       - Authentication
 *     summary: Authenticates a user.
 *     requestBody:
 *       description: Login credentials.
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             $ref: '#/components/schemas/LoginRequest'
 *     responses:
 *       200:
 *         description: User authenticated successfully.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 status:
 *                   type: string
 *                   example: "success"
 *                 message:
 *                   type: string
 *                   example: "Authenticated successfully."
 *                 token:
 *                   type: string
 *                   example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
 *       401:
 *         description: Invalid credentials.
 */
app.post('/login', (req, res) => {
  const { username, password } = req.body;
  // Dummy authentication; replace with real logic.
  if (username === 'admin' && password === 'password') {
    res.json({
      status: "success",
      message: "Authenticated successfully",
      token: "fake-jwt-token"
    });
  } else {
    res.status(401).json({ detail: "Invalid credentials" });
  }
});

/**
 * @swagger
 * /health:
 *   get:
 *     tags:
 *       - Health
 *     summary: Health check endpoint.
 *     responses:
 *       200:
 *         description: Service is running.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 status:
 *                   type: string
 *                   example: "running"
 *                 timestamp:
 *                   type: string
 *                   example: "2024-04-27T12:34:56Z"
 */
app.get('/health', (req, res) => {
  res.json({ status: "running", timestamp: new Date().toISOString() });
});

/**
 * @swagger
 * /generate_report:
 *   post:
 *     tags:
 *       - Reporting
 *     summary: Generate a report.
 *     requestBody:
 *       description: Report generation parameters.
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               report_type:
 *                 type: string
 *                 example: "sales_summary"
 *     responses:
 *       200:
 *         description: Report generated successfully.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 status:
 *                   type: string
 *                   example: "success"
 *                 report:
 *                   type: object
 *       500:
 *         description: Report generation failed.
 */
app.post('/generate_report', (req, res) => {
  try {
    const { report_type } = req.body;
    const report = generateReport(report_type);
    res.json({ status: "success", report: report });
  } catch (error) {
    logger.error("Report generation failed: " + error.stack);
    res.status(500).json({ detail: "Report generation failed: " + error.message });
  }
});

/**
 * @swagger
 * /send_newsletter:
 *   post:
 *     tags:
 *       - Newsletter
 *     summary: Send a newsletter.
 *     requestBody:
 *       description: Newsletter content.
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               content:
 *                 type: string
 *                 example: "This is the latest newsletter update..."
 *     responses:
 *       200:
 *         description: Newsletter sent successfully.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 status:
 *                   type: string
 *                   example: "success"
 *                 result:
 *                   type: object
 *       500:
 *         description: Newsletter sending failed.
 */
app.post('/send_newsletter', (req, res) => {
  try {
    const { content } = req.body;
    const result = sendNewsletter(content);
    res.json({ status: "success", result: result });
  } catch (error) {
    logger.error("Newsletter sending failed: " + error.stack);
    res.status(500).json({ detail: "Newsletter sending failed: " + error.message });
  }
});

/**
 * @swagger
 * /agent_chat:
 *   post:
 *     tags:
 *       - Chat Bot
 *     summary: Process a chat interaction via the RLG Agent Chat Bot.
 *     requestBody:
 *       description: Chat message payload.
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               message:
 *                 type: string
 *                 example: "What are today's insights?"
 *     responses:
 *       200:
 *         description: Chat response generated successfully.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 status:
 *                   type: string
 *                   example: "success"
 *                 reply:
 *                   type: object
 *       500:
 *         description: Chat processing failed.
 */
app.post('/agent_chat', (req, res) => {
  try {
    const { message } = req.body;
    const reply = processAgentChat(message);
    res.json({ status: "success", reply: reply });
  } catch (error) {
    logger.error("Agent chat failed: " + error.stack);
    res.status(500).json({ detail: "Chat processing failed: " + error.message });
  }
});

// ---------------------------------------------------------------------
// Additional Endpoint: Dynamic Pricing based on Geolocation
// ---------------------------------------------------------------------
/**
 * @swagger
 * /pricing:
 *   get:
 *     tags:
 *       - Pricing
 *     summary: Retrieve dynamic pricing based on user's IP geolocation.
 *     responses:
 *       200:
 *         description: Dynamic pricing details returned.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 region:
 *                   type: string
 *                   example: "Special Region"
 *                 pricing:
 *                   type: object
 *                   properties:
 *                     Creator:
 *                       type: object
 *                       properties:
 *                         weekly:
 *                           type: number
 *                           example: 35
 *                         monthly:
 *                           type: number
 *                           example: 99
 *                     Pro:
 *                       type: object
 *                       properties:
 *                         weekly:
 *                           type: number
 *                           example: 65
 *                         monthly:
 *                           type: number
 *                           example: 199
 *                     Enterprise:
 *                       type: object
 *                       properties:
 *                         monthly:
 *                           type: number
 *                           example: 699
 *                     "RLG Media Pack":
 *                       type: object
 *                       properties:
 *                         monthly:
 *                           type: number
 *                           example: 2500
 *                 special_message:
 *                   type: string
 *                   example: "עם ישראל חי!, הפתרון הטכנולוגי שישנה את העתיד."
 */
app.get('/pricing', (req, res) => {
  // In a real implementation, use a geolocation service to get user country from req.ip
  const userIp = req.ip || "";
  const pricingData = getPricing(userIp);
  res.json(pricingData);
});

// ---------------------------------------------------------------------
// Start the Server via Uvicorn equivalent or Node listener
// ---------------------------------------------------------------------
const PORT = process.env.PORT || 8000;
app.listen(PORT, () => {
  console.log(`RLG Data & RLG Fans API is running on port ${PORT}`);
});